refactor(aero): move wing outline to a separate util
